+++ crates/icydb-core/src/db/mod.rs
@@ -21,6 +21,7 @@ pub(in crate::db) mod data;
 pub(in crate::db) mod direction;
 pub(in crate::db) mod executor;
 pub(in crate::db) mod index;
+pub(in crate::db) mod numeric;
 pub(in crate::db) mod relation;
 
 use crate::{